move atomic backend and consensus callbacks
