+++ consensus/dummy/consensus.go
@@ -59,6 +59,7 @@ type (
 		block *types.Block,
 		parent *types.Header,
 		statedb *state.StateDB,
+		config *params.ChainConfig,
 	) (
 		blockFeeContribution *big.Int,
 		extDataGasUsed *big.Int,
@@ -371,7 +372,7 @@ func (eng *DummyEngine) Finalize(chain consensus.ChainHeaderReader, block *types
 		err                          error
 	)
 	if eng.cb.OnExtraStateChange != nil {
-		contribution, extDataGasUsed, err = eng.cb.OnExtraStateChange(block, parent, state)
+		contribution, extDataGasUsed, err = eng.cb.OnExtraStateChange(block, parent, state, chain.Config())
 		if err != nil {
 			return err
 		}
