+++ plugin/evm/atomic/vm/bonus_blocks.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
-package evm
+package vm
 
 import "github.com/ava-labs/avalanchego/ids"
