+++ plugin/evm/extension/config.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/codec"
 	"github.com/ava-labs/avalanchego/database"
+	"github.com/ava-labs/avalanchego/database/versiondb"
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow/consensus/snowman"
 	avalanchecommon "github.com/ava-labs/avalanchego/snow/engine/common"
@@ -16,6 +17,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/snow/engine/snowman/block"
 
+	"github.com/ava-labs/coreth/consensus/dummy"
 	"github.com/ava-labs/coreth/eth"
 	"github.com/ava-labs/coreth/params/extras"
 	"github.com/ava-labs/coreth/plugin/evm/config"
@@ -23,6 +25,7 @@ import (
 	"github.com/ava-labs/coreth/plugin/evm/sync"
 	"github.com/ava-labs/coreth/sync/handlers"
 
+	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/core/types"
 )
 
@@ -50,6 +53,10 @@ type ExtensibleVM interface {
 	Ethereum() *eth.Ethereum
 	// Config returns the configuration for the VM
 	Config() config.Config
+	// ReadLastAccepted returns the last accepted block hash and height
+	ReadLastAccepted() (common.Hash, uint64, error)
+	// VersionDB returns the versioned database for the VM
+	VersionDB() *versiondb.Database
 	// SyncerClient returns the syncer client for the VM
 	SyncerClient() sync.Client
 }
@@ -113,6 +120,10 @@ type Config struct {
 	// for encoding and decoding network messages.
 	// It's required and should be non-nil
 	NetworkCodec codec.Manager
+	// ConsensusCallbacks is the consensus callbacks to use
+	// for the VM to be used in consensus engine.
+	// Callback functions can be nil.
+	ConsensusCallbacks dummy.ConsensusCallbacks
 	// SyncSummaryProvider is the sync summary provider to use
 	// for the VM to be used in syncer.
 	// It's required and should be non-nil
