+++ plugin/evm/vm.go
@@ -23,14 +23,12 @@ import (
 	"github.com/ava-labs/avalanchego/network/p2p/gossip"
 	avalanchegoConstants "github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/coreth/network"
-	"github.com/ava-labs/coreth/plugin/evm/customtypes"
 	"github.com/ava-labs/coreth/plugin/evm/extension"
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/ava-labs/coreth/consensus/dummy"
 	"github.com/ava-labs/coreth/constants"
 	"github.com/ava-labs/coreth/core"
-	"github.com/ava-labs/coreth/core/state"
 	"github.com/ava-labs/coreth/core/txpool"
 	"github.com/ava-labs/coreth/eth"
 	"github.com/ava-labs/coreth/eth/ethconfig"
@@ -40,7 +38,6 @@ import (
 	"github.com/ava-labs/coreth/params"
 	"github.com/ava-labs/coreth/params/extras"
 	"github.com/ava-labs/coreth/plugin/evm/atomic"
-	atomicstate "github.com/ava-labs/coreth/plugin/evm/atomic/state"
 	atomictxpool "github.com/ava-labs/coreth/plugin/evm/atomic/txpool"
 	atomicvm "github.com/ava-labs/coreth/plugin/evm/atomic/vm"
 	"github.com/ava-labs/coreth/plugin/evm/config"
@@ -51,7 +48,6 @@ import (
 	"github.com/ava-labs/coreth/plugin/evm/upgrade/acp176"
 	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap5"
 	"github.com/ava-labs/coreth/triedb/hashdb"
-	"github.com/ava-labs/coreth/utils"
 
 	"github.com/ava-labs/libevm/core/rawdb"
 	"github.com/ava-labs/libevm/core/types"
@@ -88,7 +84,6 @@ import (
 
 	"github.com/ava-labs/avalanchego/cache/lru"
 	"github.com/ava-labs/avalanchego/codec"
-	"github.com/ava-labs/avalanchego/codec/linearcodec"
 	"github.com/ava-labs/avalanchego/database"
 	"github.com/ava-labs/avalanchego/database/versiondb"
 	"github.com/ava-labs/avalanchego/ids"
@@ -97,7 +92,6 @@ import (
 	"github.com/ava-labs/avalanchego/snow/engine/snowman/block"
 	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/avalanchego/utils/formatting/address"
-	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/perms"
 	"github.com/ava-labs/avalanchego/utils/profiler"
 	"github.com/ava-labs/avalanchego/utils/set"
@@ -119,7 +113,7 @@ var (
 	_ block.BuildBlockWithContextChainVM = (*VM)(nil)
 	_ block.StateSyncableVM              = (*VM)(nil)
 	_ statesyncclient.EthBlockParser     = (*VM)(nil)
-	_ secp256k1fx.VM                     = (*VM)(nil)
+	_ vmsync.BlockAcceptor               = (*VM)(nil)
 )
 
 const (
@@ -180,7 +174,6 @@ var (
 	errInvalidBlock                  = errors.New("invalid block")
 	errInvalidNonce                  = errors.New("invalid nonce")
 	errUnclesUnsupported             = errors.New("uncles unsupported")
-	errRejectedParent                = errors.New("rejected parent")
 	errNilBaseFeeApricotPhase3       = errors.New("nil base fee is invalid after apricotPhase3")
 	errNilBlockGasCostApricotPhase4  = errors.New("nil blockGasCost is invalid after apricotPhase4")
 	errInvalidHeaderPredicateResults = errors.New("invalid header predicate results")
@@ -263,24 +256,13 @@ type VM struct {
 
 	toEngine chan<- commonEng.Message
 
-	// [atomicTxRepository] maintains two indexes on accepted atomic txs.
-	// - txID to accepted atomic tx
-	// - block height to list of atomic txs accepted on block at that height
-	atomicTxRepository *atomicstate.AtomicRepository
-	// [atomicBackend] abstracts verification and processing of atomic transactions
-	atomicBackend *atomicstate.AtomicBackend
-
 	builder *blockBuilder
 
-	baseCodec codec.Registry
-	clock     *mockable.Clock
-	mempool   *atomictxpool.Mempool
+	clock   *mockable.Clock
+	mempool *atomictxpool.Mempool
 
 	shutdownChan chan struct{}
 	shutdownWg   sync.WaitGroup
-
-	fx        secp256k1fx.Fx
-	secpCache *secp256k1.RecoverCache
 
 	// Continuous Profiler
 	profiler profiler.ContinuousProfiler
@@ -315,15 +297,6 @@ type VM struct {
 	// RPC handlers (should be stopped before closing chaindb)
 	rpcHandlers []interface{ Stop() }
 }
-
-// CodecRegistry implements the secp256k1fx interface
-func (vm *VM) CodecRegistry() codec.Registry { return vm.baseCodec }
-
-// Clock implements the secp256k1fx interface
-func (vm *VM) Clock() *mockable.Clock { return vm.clock }
-
-// Logger implements the secp256k1fx interface
-func (vm *VM) Logger() logging.Logger { return vm.ctx.Log }
 
 // Initialize implements the snowman.ChainVM interface
 func (vm *VM) Initialize(
@@ -428,7 +401,7 @@ func (vm *VM) Initialize(
 	vm.ethConfig.Genesis = g
 	vm.ethConfig.NetworkId = vm.chainID.Uint64()
 	vm.genesisHash = vm.ethConfig.Genesis.ToBlock().Hash() // must create genesis hash before [vm.readLastAccepted]
-	lastAcceptedHash, lastAcceptedHeight, err := vm.readLastAccepted()
+	lastAcceptedHash, lastAcceptedHeight, err := vm.ReadLastAccepted()
 	if err != nil {
 		return err
 	}
@@ -491,7 +464,6 @@ func (vm *VM) Initialize(
 	}
 
 	vm.chainConfig = g.Config
-	vm.secpCache = secp256k1.NewRecoverCache(secpCacheSize)
 
 	// TODO: read size from settings
 	vm.mempool, err = atomictxpool.NewMempool(chainCtx, vm.sdkMetrics, defaultMempoolSize, vm.verifyTxAtTip)
@@ -541,41 +513,8 @@ func (vm *VM) Initialize(
 	if err := vm.initializeChain(lastAcceptedHash); err != nil {
 		return err
 	}
-	// initialize bonus blocks on mainnet
-	var (
-		bonusBlockHeights map[uint64]ids.ID
-	)
-	if vm.ctx.NetworkID == avalanchegoConstants.MainnetID {
-		bonusBlockHeights, err = readMainnetBonusBlocks()
-		if err != nil {
-			return fmt.Errorf("failed to read mainnet bonus blocks: %w", err)
-		}
-	}
-
-	// initialize atomic repository
-	vm.atomicTxRepository, err = atomicstate.NewAtomicTxRepository(vm.versiondb, atomic.Codec, lastAcceptedHeight)
-	if err != nil {
-		return fmt.Errorf("failed to create atomic repository: %w", err)
-	}
-	vm.atomicBackend, err = atomicstate.NewAtomicBackend(
-		vm.ctx.SharedMemory, bonusBlockHeights,
-		vm.atomicTxRepository, lastAcceptedHeight, lastAcceptedHash,
-		vm.config.CommitInterval,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to create atomic backend: %w", err)
-	}
 
 	go vm.ctx.Log.RecoverAndPanic(vm.startContinuousProfiler)
-
-	// so [vm.baseCodec] is a dummy codec use to fulfill the secp256k1fx VM
-	// interface. The fx will register all of its types, which can be safely
-	// ignored by the VM's codec.
-	vm.baseCodec = linearcodec.NewDefault()
-
-	if err := vm.fx.Initialize(vm); err != nil {
-		return err
-	}
 
 	// Add p2p warp message warpHandler
 	warpHandler := acp118.NewCachedHandler(meteredCache, vm.warpBackend, vm.ctx.WarpSigner)
@@ -635,7 +574,6 @@ func (vm *VM) initializeChain(lastAcceptedHash common.Hash) error {
 	if err != nil {
 		return err
 	}
-	callbacks := vm.createConsensusCallbacks()
 
 	// If the gas target is specified, calculate the desired target excess and
 	// use it during block creation.
@@ -653,7 +591,7 @@ func (vm *VM) initializeChain(lastAcceptedHash common.Hash) error {
 		eth.Settings{MaxBlocksPerRequest: vm.config.MaxBlocksPerRequest},
 		lastAcceptedHash,
 		dummy.NewDummyEngine(
-			callbacks,
+			vm.extensionConfig.ConsensusCallbacks,
 			dummy.Mode{},
 			vm.clock,
 			desiredTargetExcess,
@@ -815,252 +753,11 @@ func (vm *VM) initChainState(lastAcceptedBlock *types.Block) error {
 	return vm.ctx.Metrics.Register(chainStateMetricsPrefix, chainStateRegisterer)
 }
 
-func (vm *VM) createConsensusCallbacks() dummy.ConsensusCallbacks {
-	return dummy.ConsensusCallbacks{
-		OnFinalizeAndAssemble: vm.onFinalizeAndAssemble,
-		OnExtraStateChange:    vm.onExtraStateChange,
-	}
-}
-
-func (vm *VM) preBatchOnFinalizeAndAssemble(header *types.Header, state *state.StateDB, txs []*types.Transaction) ([]byte, *big.Int, *big.Int, error) {
-	for {
-		tx, exists := vm.mempool.NextTx()
-		if !exists {
-			break
-		}
-		// Take a snapshot of [state] before calling verifyTx so that if the transaction fails verification
-		// we can revert to [snapshot].
-		// Note: snapshot is taken inside the loop because you cannot revert to the same snapshot more than
-		// once.
-		snapshot := state.Snapshot()
-		rules := vm.rules(header.Number, header.Time)
-		if err := vm.verifyTx(tx, header.ParentHash, header.BaseFee, state, rules); err != nil {
-			// Discard the transaction from the mempool on failed verification.
-			log.Debug("discarding tx from mempool on failed verification", "txID", tx.ID(), "err", err)
-			vm.mempool.DiscardCurrentTx(tx.ID())
-			state.RevertToSnapshot(snapshot)
-			continue
-		}
-
-		atomicTxBytes, err := atomic.Codec.Marshal(atomic.CodecVersion, tx)
-		if err != nil {
-			// Discard the transaction from the mempool and error if the transaction
-			// cannot be marshalled. This should never happen.
-			log.Debug("discarding tx due to unmarshal err", "txID", tx.ID(), "err", err)
-			vm.mempool.DiscardCurrentTx(tx.ID())
-			return nil, nil, nil, fmt.Errorf("failed to marshal atomic transaction %s due to %w", tx.ID(), err)
-		}
-		var contribution, gasUsed *big.Int
-		if rules.IsApricotPhase4 {
-			contribution, gasUsed, err = tx.BlockFeeContribution(rules.IsApricotPhase5, vm.ctx.AVAXAssetID, header.BaseFee)
-			if err != nil {
-				return nil, nil, nil, err
-			}
-		}
-		return atomicTxBytes, contribution, gasUsed, nil
-	}
-
-	if len(txs) == 0 {
-		// this could happen due to the async logic of geth tx pool
-		return nil, nil, nil, atomicvm.ErrEmptyBlock
-	}
-
-	return nil, nil, nil, nil
-}
-
-// assumes that we are in at least Apricot Phase 5.
-func (vm *VM) postBatchOnFinalizeAndAssemble(
-	header *types.Header,
-	parent *types.Header,
-	state *state.StateDB,
-	txs []*types.Transaction,
-) ([]byte, *big.Int, *big.Int, error) {
-	var (
-		batchAtomicTxs    []*atomic.Tx
-		batchAtomicUTXOs  set.Set[ids.ID]
-		batchContribution *big.Int = new(big.Int).Set(common.Big0)
-		batchGasUsed      *big.Int = new(big.Int).Set(common.Big0)
-		rules                      = vm.rules(header.Number, header.Time)
-		size              int
-	)
-
-	atomicGasLimit, err := customheader.RemainingAtomicGasCapacity(vm.chainConfigExtra(), parent, header)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	for {
-		tx, exists := vm.mempool.NextTx()
-		if !exists {
-			break
-		}
-
-		// Ensure that adding [tx] to the block will not exceed the block size soft limit.
-		txSize := len(tx.SignedBytes())
-		if size+txSize > targetAtomicTxsSize {
-			vm.mempool.CancelCurrentTx(tx.ID())
-			break
-		}
-
-		var (
-			txGasUsed, txContribution *big.Int
-			err                       error
-		)
-
-		// Note: we do not need to check if we are in at least ApricotPhase4 here because
-		// we assume that this function will only be called when the block is in at least
-		// ApricotPhase5.
-		txContribution, txGasUsed, err = tx.BlockFeeContribution(true, vm.ctx.AVAXAssetID, header.BaseFee)
-		if err != nil {
-			return nil, nil, nil, err
-		}
-		// ensure `gasUsed + batchGasUsed` doesn't exceed `atomicGasLimit`
-		if totalGasUsed := new(big.Int).Add(batchGasUsed, txGasUsed); !utils.BigLessOrEqualUint64(totalGasUsed, atomicGasLimit) {
-			// Send [tx] back to the mempool's tx heap.
-			vm.mempool.CancelCurrentTx(tx.ID())
-			break
-		}
-
-		if batchAtomicUTXOs.Overlaps(tx.InputUTXOs()) {
-			// Discard the transaction from the mempool since it will fail verification
-			// after this block has been accepted.
-			// Note: if the proposed block is not accepted, the transaction may still be
-			// valid, but we discard it early here based on the assumption that the proposed
-			// block will most likely be accepted.
-			// Discard the transaction from the mempool on failed verification.
-			log.Debug("discarding tx due to overlapping input utxos", "txID", tx.ID())
-			vm.mempool.DiscardCurrentTx(tx.ID())
-			continue
-		}
-
-		snapshot := state.Snapshot()
-		if err := vm.verifyTx(tx, header.ParentHash, header.BaseFee, state, rules); err != nil {
-			// Discard the transaction from the mempool and reset the state to [snapshot]
-			// if it fails verification here.
-			// Note: prior to this point, we have not modified [state] so there is no need to
-			// revert to a snapshot if we discard the transaction prior to this point.
-			log.Debug("discarding tx from mempool due to failed verification", "txID", tx.ID(), "err", err)
-			vm.mempool.DiscardCurrentTx(tx.ID())
-			state.RevertToSnapshot(snapshot)
-			continue
-		}
-
-		batchAtomicTxs = append(batchAtomicTxs, tx)
-		batchAtomicUTXOs.Union(tx.InputUTXOs())
-		// Add the [txGasUsed] to the [batchGasUsed] when the [tx] has passed verification
-		batchGasUsed.Add(batchGasUsed, txGasUsed)
-		batchContribution.Add(batchContribution, txContribution)
-		size += txSize
-	}
-
-	// If there is a non-zero number of transactions, marshal them and return the byte slice
-	// for the block's extra data along with the contribution and gas used.
-	if len(batchAtomicTxs) > 0 {
-		atomicTxBytes, err := atomic.Codec.Marshal(atomic.CodecVersion, batchAtomicTxs)
-		if err != nil {
-			// If we fail to marshal the batch of atomic transactions for any reason,
-			// discard the entire set of current transactions.
-			log.Debug("discarding txs due to error marshaling atomic transactions", "err", err)
-			vm.mempool.DiscardCurrentTxs()
-			return nil, nil, nil, fmt.Errorf("failed to marshal batch of atomic transactions due to %w", err)
-		}
-		return atomicTxBytes, batchContribution, batchGasUsed, nil
-	}
-
-	// If there are no regular transactions and there were also no atomic transactions to be included,
-	// then the block is empty and should be considered invalid.
-	if len(txs) == 0 {
-		// this could happen due to the async logic of geth tx pool
-		return nil, nil, nil, atomicvm.ErrEmptyBlock
-	}
-
-	// If there are no atomic transactions, but there is a non-zero number of regular transactions, then
-	// we return a nil slice with no contribution from the atomic transactions and a nil error.
-	return nil, nil, nil, nil
-}
-
-func (vm *VM) onFinalizeAndAssemble(
-	header *types.Header,
-	parent *types.Header,
-	state *state.StateDB,
-	txs []*types.Transaction,
-) ([]byte, *big.Int, *big.Int, error) {
-	if !vm.chainConfigExtra().IsApricotPhase5(header.Time) {
-		return vm.preBatchOnFinalizeAndAssemble(header, state, txs)
-	}
-	return vm.postBatchOnFinalizeAndAssemble(header, parent, state, txs)
-}
-
-func (vm *VM) onExtraStateChange(block *types.Block, parent *types.Header, state *state.StateDB) (*big.Int, *big.Int, error) {
-	var (
-		batchContribution *big.Int = big.NewInt(0)
-		batchGasUsed      *big.Int = big.NewInt(0)
-		header                     = block.Header()
-		rules                      = vm.rules(header.Number, header.Time)
-	)
-
-	txs, err := atomic.ExtractAtomicTxs(customtypes.BlockExtData(block), rules.IsApricotPhase5, atomic.Codec)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// If [atomicBackend] is nil, the VM is still initializing and is reprocessing accepted blocks.
-	if vm.atomicBackend != nil {
-		if vm.atomicBackend.IsBonus(block.NumberU64(), block.Hash()) {
-			log.Info("skipping atomic tx verification on bonus block", "block", block.Hash())
-		} else {
-			// Verify [txs] do not conflict with themselves or ancestor blocks.
-			if err := vm.verifyTxs(txs, block.ParentHash(), block.BaseFee(), block.NumberU64(), rules); err != nil {
-				return nil, nil, err
-			}
-		}
-		// Update the atomic backend with [txs] from this block.
-		//
-		// Note: The atomic trie canonically contains the duplicate operations
-		// from any bonus blocks.
-		_, err := vm.atomicBackend.InsertTxs(block.Hash(), block.NumberU64(), block.ParentHash(), txs)
-		if err != nil {
-			return nil, nil, err
-		}
-	}
-
-	// If there are no transactions, we can return early.
-	if len(txs) == 0 {
-		return nil, nil, nil
-	}
-
-	for _, tx := range txs {
-		if err := tx.UnsignedAtomicTx.EVMStateTransfer(vm.ctx, state); err != nil {
-			return nil, nil, err
-		}
-		// If ApricotPhase4 is enabled, calculate the block fee contribution
-		if rules.IsApricotPhase4 {
-			contribution, gasUsed, err := tx.BlockFeeContribution(rules.IsApricotPhase5, vm.ctx.AVAXAssetID, block.BaseFee())
-			if err != nil {
-				return nil, nil, err
-			}
-
-			batchContribution.Add(batchContribution, contribution)
-			batchGasUsed.Add(batchGasUsed, gasUsed)
-		}
-	}
-
-	// If ApricotPhase5 is enabled, enforce that the atomic gas used does not exceed the
-	// atomic gas limit.
-	if rules.IsApricotPhase5 {
-		atomicGasLimit, err := customheader.RemainingAtomicGasCapacity(vm.chainConfigExtra(), parent, header)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if !utils.BigLessOrEqualUint64(batchGasUsed, atomicGasLimit) {
-			return nil, nil, fmt.Errorf("atomic gas used (%d) by block (%s), exceeds atomic gas limit (%d)", batchGasUsed, block.Hash().Hex(), atomicGasLimit)
-		}
-	}
-	return batchContribution, batchGasUsed, nil
-}
-
 func (vm *VM) SetState(_ context.Context, state snow.State) error {
+	if err := vm.atomicVM.SetState(nil, state); err != nil {
+		return fmt.Errorf("failed to set atomic VM state: %w", err)
+	}
+
 	switch state {
 	case snow.StateSyncing:
 		vm.bootstrapped.Set(false)
@@ -1087,8 +784,7 @@ func (vm *VM) onBootstrapStarted() error {
 	// Ensure snapshots are initialized before bootstrapping (i.e., if state sync is skipped).
 	// Note calling this function has no effect if snapshots are already initialized.
 	vm.blockChain.InitializeSnapshots()
-
-	return vm.fx.Bootstrapping()
+	return nil
 }
 
 // onNormalOperationsStarted marks this VM as bootstrapped
@@ -1097,9 +793,6 @@ func (vm *VM) onNormalOperationsStarted() error {
 		return nil
 	}
 	vm.bootstrapped.Set(true)
-	if err := vm.fx.Bootstrapped(); err != nil {
-		return err
-	}
 	// Initialize goroutines related to block building
 	// once we enter normal operation as there is no need to handle mempool gossip before this point.
 	return vm.initBlockBuilding()
@@ -1534,7 +1227,7 @@ func (vm *VM) CreateHandlers(context.Context) (map[string]http.Handler, error) {
 // getAtomicTx returns the requested transaction, status, and height.
 // If the status is Unknown, then the returned transaction will be nil.
 func (vm *VM) getAtomicTx(txID ids.ID) (*atomic.Tx, atomic.Status, uint64, error) {
-	if tx, height, err := vm.atomicTxRepository.GetByTxID(txID); err == nil {
+	if tx, height, err := vm.atomicVM.AtomicTxRepository.GetByTxID(txID); err == nil {
 		return tx, atomic.Accepted, height, nil
 	} else if err != database.ErrNotFound {
 		return nil, atomic.Unknown, 0, err
@@ -1610,84 +1303,7 @@ func (vm *VM) verifyTxAtTip(tx *atomic.Tx) error {
 
 	// We don’t need to revert the state here in case verifyTx errors, because
 	// [preferredState] is thrown away either way.
-	return vm.verifyTx(tx, parentHeader.Hash(), nextBaseFee, preferredState, rules)
-}
-
-// verifyTx verifies that [tx] is valid to be issued into a block with parent block [parentHash]
-// and validated at [state] using [rules] as the current rule set.
-// Note: verifyTx may modify [state]. If [state] needs to be properly maintained, the caller is responsible
-// for reverting to the correct snapshot after calling this function. If this function is called with a
-// throwaway state, then this is not necessary.
-func (vm *VM) verifyTx(tx *atomic.Tx, parentHash common.Hash, baseFee *big.Int, state *state.StateDB, rules extras.Rules) error {
-	parent, err := vm.GetExtendedBlock(context.TODO(), ids.ID(parentHash))
-	if err != nil {
-		return fmt.Errorf("failed to get parent block: %w", err)
-	}
-	atomicBackend := &atomicvm.VerifierBackend{
-		Ctx:          vm.ctx,
-		Fx:           &vm.fx,
-		Rules:        rules,
-		Bootstrapped: vm.bootstrapped.Get(),
-		BlockFetcher: vm,
-		SecpCache:    vm.secpCache,
-	}
-	if err := tx.UnsignedAtomicTx.Visit(&atomicvm.SemanticVerifier{
-		Backend: atomicBackend,
-		Tx:      tx,
-		Parent:  parent,
-		BaseFee: baseFee,
-	}); err != nil {
-		return err
-	}
-	return tx.UnsignedAtomicTx.EVMStateTransfer(vm.ctx, state)
-}
-
-// verifyTxs verifies that [txs] are valid to be issued into a block with parent block [parentHash]
-// using [rules] as the current rule set.
-func (vm *VM) verifyTxs(txs []*atomic.Tx, parentHash common.Hash, baseFee *big.Int, height uint64, rules extras.Rules) error {
-	// Ensure that the parent was verified and inserted correctly.
-	if !vm.blockChain.HasBlock(parentHash, height-1) {
-		return errRejectedParent
-	}
-
-	ancestorID := ids.ID(parentHash)
-	// If the ancestor is unknown, then the parent failed verification when
-	// it was called.
-	// If the ancestor is rejected, then this block shouldn't be inserted
-	// into the canonical chain because the parent will be missing.
-	ancestor, err := vm.GetExtendedBlock(context.TODO(), ancestorID)
-	if err != nil {
-		return errRejectedParent
-	}
-
-	// Ensure each tx in [txs] doesn't conflict with any other atomic tx in
-	// a processing ancestor block.
-	inputs := set.Set[ids.ID]{}
-	atomicBackend := &atomicvm.VerifierBackend{
-		Ctx:          vm.ctx,
-		Fx:           &vm.fx,
-		Rules:        rules,
-		Bootstrapped: vm.bootstrapped.Get(),
-		BlockFetcher: vm,
-		SecpCache:    vm.secpCache,
-	}
-	for _, atomicTx := range txs {
-		utx := atomicTx.UnsignedAtomicTx
-		if err := utx.Visit(&atomicvm.SemanticVerifier{
-			Backend: atomicBackend,
-			Tx:      atomicTx,
-			Parent:  ancestor,
-			BaseFee: baseFee,
-		}); err != nil {
-			return fmt.Errorf("invalid block due to failed semanatic verify: %w at height %d", err, height)
-		}
-		txInputs := utx.InputUTXOs()
-		if inputs.Overlaps(txInputs) {
-			return atomicvm.ErrConflictingAtomicInputs
-		}
-		inputs.Union(txInputs)
-	}
-	return nil
+	return vm.atomicVM.VerifyTx(tx, parentHeader.Hash(), nextBaseFee, preferredState, rules)
 }
 
 // GetAtomicUTXOs returns the utxos that at least one of the provided addresses is
@@ -1771,7 +1387,7 @@ func (vm *VM) startContinuousProfiler() {
 // last accepted block hash and height by reading directly from [vm.chaindb] instead of relying
 // on [chain].
 // Note: assumes [vm.chaindb] and [vm.genesisHash] have been initialized.
-func (vm *VM) readLastAccepted() (common.Hash, uint64, error) {
+func (vm *VM) ReadLastAccepted() (common.Hash, uint64, error) {
 	// Attempt to load last accepted block to determine if it is necessary to
 	// initialize state with the genesis block.
 	lastAcceptedBytes, lastAcceptedErr := vm.acceptedBlockDB.Get(lastAcceptedKey)
